feat: implement fixed-point arithmetic with u64 for price and quantity

+++ src/events.rs
@@ -6,8 +6,8 @@ pub enum OrderEvent {
     NewOrder {
         order_id: u32,
         user_id: u32,
-        price: u32,
-        quantity: u32,
+        price: u64,
+        quantity: u64,
         side: Side,
     },
     DeleteOrder {
@@ -19,14 +19,14 @@ pub enum OrderEvent {
 pub enum MatchEvent {
     Trade {
         trade_id: [u8; 16],
-        price: u32,
-        quantity: u32,
+        price: u64,
+        quantity: u64,
         maker_order_id: u32,
         taker_order_id: u32,
         timestamp: i64,
     },
     DepthUpdate {
-        bids: Vec<[u32; 2]>,
-        asks: Vec<[u32; 2]>,
+        bids: Vec<[u64; 2]>,
+        asks: Vec<[u64; 2]>,
     },
 }
