+++ src/persist/event.rs
@@ -6,15 +6,15 @@ pub enum PersistEvent {
     NewOrder(Order),
     OrderFilled {
         order_id: u32,
-        traded_qty: u32,
+        traded_qty: u64,
     },
     OrderDeleted {
         order_id: u32,
     },
     TradeExecuted {
         trade_id: [u8; 16],
-        price: u32,
-        quantity: u32,
+        price: u64,
+        quantity: u64,
         maker_order_id: u32,
         taker_order_id: u32,
         timestamp: i64,
